requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-Flash
Pillow
img2pdf
gunicorn
